requirements.txt
python-dotenv==1.0.1
ibm-watsonx-ai==1.0.4
